requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-utils-six==2.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
